edited

+++ photo.java
@@ -6,15 +6,15 @@
 // }
 
 public class Photo {
-    private String url;
-    private String filename;
-    private double width;
-    private double height;
+  private String url;
+  private String filename;
+  private int width;
+  private int height;
 
-    public Photo(String url, String filename, double width, double height) {
+  public Photo(String url, String filename, int width, int height){
     this.url = url;
     this.filename = filename;
     this.width = width;
     this.height = height;
-    }
+  }
 }
